% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{adj_R2}
\alias{adj_R2}
\title{Calculate the adjusted R square}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{A vector of predicted numerical values}

\item{truth}{A vector of observed numerical values}

\item{n_p}{The number of parameters to estimate except the intercept}
}
\value{
A numerical value
}
\description{
Calculate the adjusted R square
}
\examples{
pred = c(1,2,3,4,5)
truth = c(1.2, 2.2, 3.0, 3.8, 5.5)
adj_R2(pred, truth, 2)
# returns
0.9293084
}
