% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{df_scale}
\alias{df_scale}
\title{Center and scale a data frame}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{a data frame}
}
\value{
A tibble
}
\description{
Center and scale a data frame
}
\examples{
x = tibble(x = c(1,2,3),
           y = c(4,5,6),
           z = c("a","b","c")
          )
df_scale(x)
# returns
      x     y z    
  <dbl> <dbl> <chr>
1    -1    -1 a    
2     0     0 b    
3     1     1 c    
}
