% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_min}
\alias{is_min}
\title{Identify the minimum and max value}
\usage{
is_min(x, na.rm=T)
is_max(x, na.rm=T)
}
\arguments{
\item{x}{A vector}
}
\value{
A logical vector
}
\description{
Identify the minimum and max value
}
\examples{
x = c(1,2,3)
is_min(x, na.rm=T)
# returns
# [1] TRUE, FALSE, FALSE
}
