% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Sum of log likelihood for a particular probability distribution}
\usage{
# Normal distribition
log_likelihood_norm(x, mean, sd)

# Uniform distribution
log_likelihood_unif(x. min, max)

# Chi square distribution
log_likelihood_chisq(x, df)

# F distribution
log_likelihood_f(x, df1, df2)

# t distribution
log_likelihood_t(x, df)
}
\arguments{
\item{x}{a vector}
}
\value{
The sum of the log likelihood
}
\description{
Sum of log likelihood for a particular probability distribution
}
\examples{
x = c(10,11,12,13)
log_likelihood_norm(x, mean=11, sd = 1)
# returns
-6.675754
}
