% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{r2}
\alias{r2}
\title{Calculate the R square}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{A vector of predicted numerical values}

\item{truth}{A vector of observed numerical values}
}
\value{
A numerical value
}
\description{
Calculate the R square
}
\examples{
pred = c(1,2,3,4,5)
truth = c(1.2, 2.2, 3.0, 3.8, 5.5)
r2(pred, truth)
# returns
0.9646542
}
