% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rando}
\alias{rando}
\title{Apply functions over a vector or dataframe (tibble)}
\usage{
rando(X, n, replace = TRUE)
}
\arguments{
\item{X}{An atomic vector, data frame, or tibble}

\item{n}{n samples will be sampled from X}

\item{replace}{default TRUE, this allows sampling with replacement}
}
\value{
vector or data frame
}
\description{
Returns a specified samples or rows from a vector or data frame.
}
\details{
This function take a vector or dataframe. If a vector is the input, this function returns a specified number of random samples from this vector. If a data frame is the imput, this function picks a specified number of random rows and returns a data frame with these rows.
}
\examples{
X = c(1,2,3)
rando(X, 2, replace = F)
}
