% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rep_mat}
\alias{rep_mat}
\title{Function to return a detaframe or matrix where each original row is replicated.}
\usage{
rep_mat(X, M=1, N=1)
}
\arguments{
\item{X}{A data frame or matrix}

\item{M}{The number of replication for the rows}

\item{N}{The number of replication for the columns}
}
\value{
A data frame or matrix
}
\description{
Function to return a detaframe or matrix where each original row is replicated.
}
\examples{
X = matrix(c(1:10), ncol = 5)
rep_mat(X, M= 2, N=2)
# This returs
#  
}
