% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{sensitivity}
\alias{sensitivity}
\alias{specificity}
\title{Calculate the sensitivity, specificity, precision, recall, accuracy, and F1 scores}
\usage{
sensitivity(pred, truth)
specificity(pred, truth)
precision(pred, truth)
recall(pred, truth)
accuracy(pred, truth)
f1(pred, truth)
}
\arguments{
\item{pred}{A vector of predicted binary or boolean values}

\item{truth}{A vector of true binary or boolean values}
}
\value{
A value
}
\description{
Calculate the sensitivity, specificity, precision, recall, accuracy, and F1 scores
}
\examples{
pred = c(rep(1,40),rep(0,10), rep(1,45),rep(0,5))
truth = c(rep(1, 50),rep(0, 50))
sensitivity(pred, truth)
# returns
0.8
}
