% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{should_stop_modified}
\alias{should_stop_modified}
\title{This stores functions for BMI510 final project}
\usage{
should_stop_modified(expr)
}
\arguments{
\item{expr}{code to evaluate}
}
\description{
20 functions are documented in this package
}
\details{
xxx
}
\examples{
should_stop(stop("Hi"))
}
\keyword{internal}
