% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rando}
\alias{rando}
\title{Apply functions over a vector or dataframe (tibble)}
\usage{
rando(X, n, replace = T)
}
\arguments{
\item{X}{An atomic vector, data frame, or tibble}

\item{n}{n samples will be sampled from X}

\item{replace}{default TRUE, this allows sampling with replacement}
}
\value{
3,2
}
\description{
Returns a specified samples or rows from a vector or data frame.
}
\details{
This stores functions for BMI510 final project

20 functions are documented in this package

rando(X, n, replace = TRUE)
}
\examples{
X = c(1,2,3)
rando(X, 2, replace = F)
}
