% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rando}
\alias{rando}
\title{xxx}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{dataframe}

\item{n}{start with 1, n, replace}
}
\description{
xxx
}
\details{
xxx
}
\examples{
rando(x, 3, replace = F)
}
